rebrand escalated daemon into herder daemon
