--- src/escalated_daemon/ipc.rs
@@ -1,8 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use crate::writer_process::ipc::WriterProcessConfig;
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SpawnWriter {
-    pub init_config: WriterProcessConfig,
-}
